move and rename android stuff
